Move thread-safe string to another file
