+++ exp/lighthorizon/index/cmd/batch/reduce/main.go
@@ -19,11 +19,6 @@ var (
 	parallel = uint32(16)
 )
 
-type SafeStringSet struct {
-	lock sync.Mutex
-	set  map[string]struct{}
-}
-
 type ReduceConfig struct {
 	JobIndex       uint32
 	MapJobCount    uint32
@@ -245,26 +240,6 @@ func main() {
 
 		wg.Wait()
 	}
-}
-
-func NewSafeStringSet() *SafeStringSet {
-	return &SafeStringSet{
-		lock: sync.Mutex{},
-		set:  map[string]struct{}{},
-	}
-}
-
-func (set *SafeStringSet) Contains(key string) bool {
-	defer set.lock.Unlock()
-	set.lock.Lock()
-	_, ok := set.set[key]
-	return ok
-}
-
-func (set *SafeStringSet) Add(key string) {
-	defer set.lock.Unlock()
-	set.lock.Lock()
-	set.set[key] = struct{}{}
 }
 
 func shouldAccountBeProcessed(account string,
